Fix: warning with (int32)
Former-commit-id: e305427233b5d628845949695ef68fa76351b67e

+++ Sources/Internal/UI/UIFileSystemDialog.cpp
@@ -575,7 +575,7 @@ void UIFileSystemDialog::HistoryButtonPressed(BaseObject *obj, void *data, void 
     }
     else if (obj == historyForwardButton)
     {
-        if(historyPosition < foldersHistory.size() - 1)
+        if(historyPosition < (int32)foldersHistory.size() - 1)
         {
             SetCurrentDir(foldersHistory[historyPosition + 1]);
         }
